Переносит часть логики в presenter
